+++ CHANGELOG.md
@@ -14,6 +14,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 + [[Backend-371]](https://www.notion.so/Talle-XS-Filtrado-de-materias-por-carrera-4e6933405e994d7b8d624d1b5ce22588?pvs=4)
 + New /subjects endpoint wit self and career_id filters.
 
++ [[Backend-373]](https://www.notion.so/Talle-S-Editar-un-tema-1e31bd9a08c746e8b80622dcc6564bb2?pvs=4)
++ Added topics update endpoint
++ Added unit tests for topics controller, update endpoint
+
++ [[Backend-370]](https://www.notion.so/Talle-M-Seleccionar-y-cambiar-carrera-de-usuario-editar-pefil-4a95c866da3c42deba0eeed8750c3614)
++ Modify profile endpoint to allow changing the user's career
++ Modify profile endpoint to show the user's career
++ Create associations between user and career
+
 + [[Backend-375]](https://www.notion.so/Talle-M-Ver-detalle-de-un-tema-b6ebcc0ae5ee4ca4aae9accbd11e7078?pvs=4)
 + Changed topics#show route from proposed_topics/:availability_id to topics/:id
 + Added the logig to topics#show